# 资源/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: Hope and Sparks API
  version: 1.0.0
  description: 'Apifox 导入版接口规范。详细业务说明、数据库落库规则和枚举映射见根目录

    `Hope_and_Sparks_API接口文档.md`。

    '
servers:
- url: http://localhost:8080
  description: 本地开发环境
- url: /
  description: 相对路径
security:
- bearerAuth: []
...
paths:
...
  /api/v1/onboarding/answers:
    post:
      tags:
      - Spark画像与首次引导
      summary: 提交单轮画像回答
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                questionId:
                  type: string
                  example: q_learning_goal
                  description: 题目 ID，对应 question_bank.id
                answer:
                  type: object
                  additionalProperties: true
                  description: 用户答案，按题型可为字符串、数组或对象
                answerText:
                  type: string
                  description: 文本答案
                selectedOptions:
                  type: array
                  items:
                    type: string
                  description: 选择题选项数组
              required:
              - questionId
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `questionId` | 是 | string | 题目 ID，对应 question_bank.id |

        | Body | `answer` | 否 | object | 用户答案，按题型可为字符串、数组或对象 |

        | Body | `answerText` | 否 | string | 文本答案 |

        | Body | `selectedOptions` | 否 | array<string> | 选择题选项数组 |


        主要数据表：user_profile、agent_memory'
...
  /api/v1/spark-profile/rebuild:
    post:
      tags:
      - Spark画像与首次引导
      summary: 重新构建画像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
                  description: 操作原因
                answers:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                  description: 批量作答内容
              required:
              - reason
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `reason` | 是 | string | 操作原因 |

        | Body | `answers` | 否 | array<object> | 批量作答内容 |


        主要数据表：user_profile、agent_memory'
...
  /api/v1/user/profile:
    put:
      tags:
      - 用户资料与个人中心
      summary: 更新个人资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                nickname:
                  type: string
                  description: 昵称，对应 sys_user.nickname
                  example: 一粒黑子
                avatar:
                  type: string
                  description: 头像地址，对应 sys_user.avatar_url
                bio:
                  type: string
                  description: 个人简介，对应 user_profile.bio
                learningDomain:
                  type: string
                  description: 学习方向，对应 user_profile.major_domain
                gradeLevel:
                  type: string
                  description: 学段/年级，对应 user_profile.grade_level
                knowledgeBaseLevel:
                  type: string
                  description: 基础水平，对应 user_profile.knowledge_base_level
                selfDiscipline:
                  type: string
                  description: 自律程度，对应 user_profile.self_discipline
                interests:
                  type: array
                  items:
                    type: string
                  description: 兴趣标签
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `nickname` | 否 | string | 昵称，对应 sys_user.nickname |

        | Body | `avatar` | 否 | string | 头像地址，对应 sys_user.avatar_url |

        | Body | `bio` | 否 | string | 个人简介，对应 user_profile.bio |

        | Body | `learningDomain` | 否 | string | 学习方向，对应 user_profile.major_domain |

        | Body | `gradeLevel` | 否 | string | 学段/年级，对应 user_profile.grade_level |

        | Body | `knowledgeBaseLevel` | 否 | string | 基础水平，对应 user_profile.knowledge_base_level |

        | Body | `selfDiscipline` | 否 | string | 自律程度，对应 user_profile.self_discipline |

        | Body | `interests` | 否 | array<string> | 兴趣标签 |


        主要数据表：sys_user、user_profile、user_settings、user_resource_favorite、blog_favorite'
...
  /api/v1/resources:
    get:
      tags:
      - 学习资源库
      summary: 获取资源列表
      parameters:
      - $ref: '#/components/parameters/type'
      - $ref: '#/components/parameters/keyword'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `keyword` | 否 | string | 搜索关键词 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：learning_resource、learning_resource_version、sys_oss_file、user_learning_record、feedback_ticket'
  /api/v1/resources/{resourceId}:
    get:
      tags:
      - 学习资源库
      summary: 获取资源详情
      parameters:
      - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `resourceId` | 是 | string | 学习资源 ID，对应 learning_resource.id |


        主要数据表：learning_resource、learning_resource_version、sys_oss_file、user_learning_record、feedback_ticket'
...
  /api/v1/videos/{videoId}:
    get:
      tags:
      - 视频学习
      summary: 获取视频详情
      parameters:
      - $ref: '#/components/parameters/videoId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `videoId` | 是 | string | 视频资源 ID，对应 learning_resource.id |


        主要数据表：learning_resource、sys_oss_file、user_learning_record'
  /api/v1/videos/{videoId}/episodes:
    get:
      tags:
      - 视频学习
      summary: 获取视频选集
      parameters:
      - $ref: '#/components/parameters/videoId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `videoId` | 是 | string | 视频资源 ID，对应 learning_resource.id |


        主要数据表：learning_resource、sys_oss_file、user_learning_record'
  /api/v1/videos/{videoId}/watch-progress:
    put:
      tags:
      - 视频学习
      summary: 上报播放进度
      parameters:
      - $ref: '#/components/parameters/videoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                positionSeconds:
                  type: integer
                  example: 120
                  description: 播放位置，单位秒
                durationSeconds:
                  type: integer
                  example: 3600
                  description: 耗时或资源总时长，单位秒
                progressPercent:
                  type: integer
                  example: 30
                  description: 学习进度百分比
                completed:
                  type: boolean
                  example: false
                  description: 是否完成
              required:
              - positionSeconds
              - progressPercent
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `videoId` | 是 | string | 视频资源 ID，对应 learning_resource.id |

        | Body | `positionSeconds` | 是 | integer | 播放位置，单位秒 |

        | Body | `durationSeconds` | 否 | integer | 耗时或资源总时长，单位秒 |

        | Body | `progressPercent` | 是 | integer | 学习进度百分比 |

        | Body | `completed` | 否 | boolean | 是否完成 |


        主要数据表：learning_resource、sys_oss_file、user_learning_record'
  /api/v1/videos/{videoId}/transcripts:
    get:
      tags:
      - 视频学习
      summary: 获取视频字幕与 AI 提示
      parameters:
      - $ref: '#/components/parameters/videoId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `videoId` | 是 | string | 视频资源 ID，对应 learning_resource.id |


        主要数据表：learning_resource、sys_oss_file、user_learning_record'
  /api/v1/documents/{documentId}:
    get:
      tags:
      - 文档阅读
      summary: 获取文档详情
      parameters:
      - $ref: '#/components/parameters/documentId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `documentId` | 是 | string | 文档 ID，对应 learning_resource.id 或 kb_document.id |


        主要数据表：learning_resource、kb_document、kb_chunk_record、user_learning_record'
  /api/v1/documents/{documentId}/outline:
    get:
      tags:
      - 文档阅读
      summary: 获取文档目录
      parameters:
      - $ref: '#/components/parameters/documentId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `documentId` | 是 | string | 文档 ID，对应 learning_resource.id 或 kb_document.id |


        主要数据表：learning_resource、kb_document、kb_chunk_record、user_learning_record'
  /api/v1/documents/{documentId}/reading-progress:
    put:
      tags:
      - 文档阅读
      summary: 更新阅读进度
      parameters:
      - $ref: '#/components/parameters/documentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                page:
                  type: integer
                  example: 12
                  description: 页码
                scrollPercent:
                  type: integer
                  example: 65
                  description: 阅读滚动百分比
                progressPercent:
                  type: integer
                  example: 40
                  description: 学习进度百分比
                completed:
                  type: boolean
                  example: false
                  description: 是否完成
              required:
              - progressPercent
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `documentId` | 是 | string | 文档 ID，对应 learning_resource.id 或 kb_document.id |

        | Body | `page` | 否 | integer | 页码 |

        | Body | `scrollPercent` | 否 | integer | 阅读滚动百分比 |

        | Body | `progressPercent` | 是 | integer | 学习进度百分比 |

        | Body | `completed` | 否 | boolean | 是否完成 |


        主要数据表：learning_resource、kb_document、kb_chunk_record、user_learning_record'
  /api/v1/documents/{documentId}/ask:
    post:
      tags:
      - 文档阅读
      summary: Sage 伴读提问
      parameters:
      - $ref: '#/components/parameters/documentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                content:
                  type: string
                  description: 正文内容
                  example: 这里是正文内容
                selectedText:
                  type: string
                  description: 用户选中的文档片段
                page:
                  type: integer
                  description: 页码
                agentId:
                  type: string
                  example: sage
                  description: 智能体 ID，如 nebula/coach/sage
              required:
              - content
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `documentId` | 是 | string | 文档 ID，对应 learning_resource.id 或 kb_document.id |

        | Body | `content` | 是 | string | 正文内容 |

        | Body | `selectedText` | 否 | string | 用户选中的文档片段 |

        | Body | `page` | 否 | integer | 页码 |

        | Body | `agentId` | 否 | string | 智能体 ID，如 nebula/coach/sage |


        主要数据表：learning_resource、kb_document、kb_chunk_record、user_learning_record'
...
  /api/v1/learning-plans/{planId}/topology/nodes/{nodeId}/resources:
    get:
      tags:
      - 学习计划日历与拓扑
      summary: 获取知识点资源网络图列表视图
      parameters:
      - $ref: '#/components/parameters/planId'
      - $ref: '#/components/parameters/nodeId'
      - $ref: '#/components/parameters/type'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `planId` | 是 | string | 学习计划 ID，对应 study_plan.id |

        | Path | `nodeId` | 是 | string | 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：study_plan、study_task、knowledge_node、knowledge_node_relation、user_knowledge_progress'
...
  /api/v1/manage/resources:
    get:
      tags:
      - Manage管理端
      summary: 管理历史生成资源
      parameters:
      - $ref: '#/components/parameters/type'
      - name: qualityStatus
        in: query
        schema:
          type: string
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Query | `type` | 否 | string | 类型筛选，具体枚举见接口文档数据字典 |

        | Query | `qualityStatus` | 否 | string | 资源质量状态筛选 |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：sys_admin、sys_role、sys_permission、feedback_ticket、kb_document、sys_operation_log、sys_agent_prompt'
  /api/v1/manage/resources/{resourceId}:
    delete:
      tags:
      - Manage管理端
      summary: 删除历史生成资源
      parameters:
      - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `resourceId` | 是 | string | 学习资源 ID，对应 learning_resource.id |


        主要数据表：sys_admin、sys_role、sys_permission、feedback_ticket、kb_document、sys_operation_log、sys_agent_prompt'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 页码，从 1 开始
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: 每页数量
    keyword:
      name: keyword
      in: query
      schema:
        type: string
      description: 搜索关键词
    type:
      name: type
      in: query
      schema:
        type: string
      description: 类型筛选，具体枚举见接口文档数据字典
...
    resourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: string
      description: 学习资源 ID，对应 learning_resource.id
...
    videoId:
      name: videoId
      in: path
      required: true
      schema:
        type: string
      description: 视频资源 ID，对应 learning_resource.id
    documentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
      description: 文档 ID，对应 learning_resource.id 或 kb_document.id
    planId:
      name: planId
      in: path
      required: true
      schema:
        type: string
      description: 学习计划 ID，对应 study_plan.id
...
    nodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: string
      description: 知识点节点 ID，对应 knowledge_node.id，兼容 knowledge_node.node_code
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    PageSuccess:
      description: 分页成功
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/PageResult'
...
  schemas:
    ApiResponse:
      type: object
      required:
      - code
      - message
      - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
          oneOf:
          - type: object
            additionalProperties: true
          - type: array
            items: {}
          - type: string
          - type: number
          - type: boolean
        requestId:
          type: string
          example: req_20260505_xxxx
    PageResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 120
        list:
          type: array
          items:
            type: object
            additionalProperties: true